Make wireless integration real
Remove the old wireless management page and replace with the new. Add
new redirect links to partners menu to sync before passing the user
along.
